requirements.txt
flask
flask-cors
firebase-admin
python-dotenv
PyJWT
flasgger 